services/payment/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-jose==3.3.0
httpx==0.27.0
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
pytest-testcontainers==3.0.0
